requirements.txt
asgiref==3.2.10
Django==3.0.8
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-model-utils==4.0.0
django-mptt==0.11.0
django-notifications-hq==1.6.0
django-password-reset==2.0
django-taggit==1.3.0
jsonfield==3.1.0
Markdown==3.2.2
mysqlclient==2.0.1
Pillow==7.2.0
Pygments==2.6.1
PyMySQL==0.9.3
pytz==2020.1
sqlparse==0.3.1
swapper==1.1.2.post1
